* Android's IAP system updated internally to use their V3 billing system for increased reliability.   - IAPManager::BuyItem() now takes an optional second parm as a user payload   - Android purchases now return full json receipt data as well   - IMPORTANT:  IAPManager::ConsumeItem() added, you must now call this after purchase if your app has consumable IAP!!  Otherwise they'll never be able to buy more than one (android only)   - Not tested all that well, so please test your apps IAP well and let me know if you notice any problems   - For existing projects, the build_prepare.bat must be modifed to copy an extra lib, check the build_prepare.bat from the updated RTAdTest example

+++ proton/RTAdTest/source/MenuTest.cpp
@@ -56,7 +56,7 @@ const string extraData("extraData");
 void sendIAPResultMessageToIAPManager(IAPManager::ResponseCode responseCode)
 {
 	Message msg(MESSAGE_CLASS_GUI, TIMER_SYSTEM, MESSAGE_TYPE_IAP_RESULT);
-	msg.SetParm1(responseCode);
+	msg.SetParm1((float)responseCode);
 	msg.SetStringParm(extraData);
 	GetApp()->m_IAPManager.OnMessage(msg);
 }
@@ -73,7 +73,7 @@ enum ItemStateCode
 void sendIAPItemStateMessageToIAPManager(ItemStateCode itemStateCode)
 {
 	Message msg(MESSAGE_CLASS_GUI, TIMER_SYSTEM, MESSAGE_TYPE_IAP_ITEM_STATE);
-	msg.SetParm1(itemStateCode);
+	msg.SetParm1((float)itemStateCode);
 	if (itemStateCode == PURCHASED)
 	{
 		msg.SetStringParm(testItem);
